+++ src/trx/game/objects/general/cutscene_player.c
@@ -1,5 +1,6 @@
 #include <trx/game/camera.h>
 #include <trx/game/collision.h>
+#include <trx/game/creature.h>
 #include <trx/game/cutscene.h>
 #include <trx/game/objects.h>
 #include <trx/game/output.h>
@@ -52,6 +53,7 @@ static void M_Control(const int16_t item_num)
 static void M_Setup(OBJECT *const obj)
 {
     obj->initialise_func = M_Initialise;
+    obj->shadow_size = (UNIT_SHADOW * 10) / 16;
     obj->control_func = M_Control;
     obj->hit_points = 1;
 }