+++ docs/CHANGELOG.md
@@ -31,6 +31,7 @@
 - added support for cross-fades to the title screen
 - added visual previews of bar colors (Graphic Options → Bars)
 - added the ability to change PS1 bar colors
+- added shadow rendering to all cutscene actors
 - added endless sprint (available previously via the `/restless` command) to the UI settings (Gameplay options → Mods → Endless sprint)
 - added endless flare time cheat (Gameplay options → Mods → Endless flare time)
 - added `O_VULTURE` for custom levels